JWT Authentication setup

+++ backend/controllers/goalController.js
@@ -2,12 +2,13 @@ const asyncHandler = require('express-async-handler')
 
 // use our goal schema
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -22,7 +23,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -31,11 +33,24 @@ const setGoal = asyncHandler(async (req, res) => {
 // @route PUT /api/goals/:id
 // @access Private
 const updateGoal = asyncHandler(async (req, res) => {
-    const goal = await Goal.findById(req.params.id)
+    const goal = await Goal.findOne({ _id: req.params.id })
 
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findOne({ _id: req.user.id })
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -47,11 +62,24 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @route DELETE /api/goals/:id
 // @access Private
 const deleteGoal = asyncHandler(async (req, res) => {
-    const goal = await Goal.findById(req.params.id)
+    const goal = await Goal.findOne({ _id: req.params.id })
 
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findOne({ _id: req.user.id })
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await Goal.deleteOne({ _id: req.params.id })
